Re-organized code for clustering spike data
This commit fixes a lot of bad coding practices in our spikesorting
module. The original TetrodeToCluster object, which does clustering
for single sessions, as well as the ClusterReportFromData object used
to plot reports, used as arguments and internal attributes an object
called dataTT. This object is returned by loadopenephys.DataSpikes
when the spike data for the session is loaded, and contains spike
timestamps, waveforms, and information such as the header and gain of
the channels. When I went to write the original MultipleSessionsToCluster
object, this messed me up. The new object, whose purpose is just to
load a bunch of sessions into one dataset and cluster it all at once,
should have been able to use all the same methods except the load_waveforms
method. However, I couldn't just return a DataSpikes object and save it, so I
re-wrote the object to store vectors of timestamps, samples, and clusters
instead. This resulted in me copying most of the TetrodeToCluster object.

In this commit, I re-wrote TetrodeToCluster and ClusterReportFromData to store
and pass vectors of timestamps, samples, and clusters instead of DataSpikes
objects.